History additions.
I took Kathy's idea and ran with it. The director now has a
copy of history. Any changes to the history field will be
made in the director. Now we need to try this with the JTable.

Adding an airport now checks to see if the airport already
exists in system. I also believe I fixed the JDialog box.
Please test this.

Signed-off-by: Brian Powell <powellb1@my.erau.edu>