Duplicates fixed
Fixed the error we were having where duplicates for time and
cost would not show in the list. Everything works as it should.
Airline filter appears to be working as well. Further testing
needed. Ready for bug fixes and final implementation of the
editRoute option as well as live updating of our table of
routes.

Signed-off-by: Brian Powell <powellb1@my.erau.edu>